Revert "move ApplicationModel"
This reverts commit 35fe758e

+++ benchmark-base/src/main/java/benchmark/service/UserService.java
@@ -8,9 +8,9 @@ public interface UserService {
 
 	public boolean createUser(User user);
 
-	public User getUser(long id);
+	public User getUser(Integer id);
 
-	public Page<User> listUser(int pageNo);
+	public Page<User> listUser(Integer pageNo);
 
 	default byte[] pingpong(byte[] object){
 		return object;
